+++ src/main/java/org/opensearch/geospatial/action/upload/geojson/UploadGeoJSONTransportAction.java
@@ -1,9 +1,15 @@
 /*
- * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ *
+ * Modifications Copyright OpenSearch Contributors. See
+ * GitHub history for details.
  */
 
-package org.opensearch.geospatial.action.geojson.upload;
+package org.opensearch.geospatial.action.upload.geojson;
 
 import org.opensearch.action.ActionListener;
 import org.opensearch.action.support.ActionFilters;
