+++ src/main/java/org/opensearch/geospatial/plugin/GeospatialPlugin.java
@@ -20,10 +20,10 @@ import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.IndexScopedSettings;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.settings.SettingsFilter;
-import org.opensearch.geospatial.action.geojson.upload.UploadGeoJSONAction;
-import org.opensearch.geospatial.action.geojson.upload.UploadGeoJSONTransportAction;
+import org.opensearch.geospatial.action.upload.geojson.UploadGeoJSONAction;
+import org.opensearch.geospatial.action.upload.geojson.UploadGeoJSONTransportAction;
 import org.opensearch.geospatial.processor.FeatureProcessor;
-import org.opensearch.geospatial.rest.action.geojson.upload.RestUploadGeoJSONAction;
+import org.opensearch.geospatial.rest.action.upload.geojson.RestUploadGeoJSONAction;
 import org.opensearch.ingest.Processor;
 import org.opensearch.plugins.ActionPlugin;
 import org.opensearch.plugins.IngestPlugin;
