+++ src/test/java/org/opensearch/geospatial/plugin/GeospatialPluginTests.java
@@ -15,9 +15,9 @@ import org.hamcrest.MatcherAssert;
 import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionResponse;
 import org.opensearch.common.settings.Settings;
-import org.opensearch.geospatial.action.geojson.upload.UploadGeoJSONAction;
+import org.opensearch.geospatial.action.upload.geojson.UploadGeoJSONAction;
 import org.opensearch.geospatial.processor.FeatureProcessor;
-import org.opensearch.geospatial.rest.action.geojson.upload.RestUploadGeoJSONAction;
+import org.opensearch.geospatial.rest.action.upload.geojson.RestUploadGeoJSONAction;
 import org.opensearch.ingest.Processor;
 import org.opensearch.plugins.ActionPlugin;
 import org.opensearch.plugins.IngestPlugin;
